pyproject.toml
[build-system]
requires = ['flit_core >=3.2,<4']
build-backend = 'flit_core.buildapi'

[project]
name = 'biodiv'
version = '0.0.1'
requires-python = '~=3.12.0'
description = 'Predicting and supporting benthic and pelagic biodiversity through geomorphometry and machine learning'
authors = [
  { name = 'Alex-Andrei Cuvuliuc' },
]
license = { file = 'LICENSE' }
readme = 'README.md'
classifiers = [
    'Programming Language :: Python :: 3',
    'License :: OSI Approved :: MIT License'
]
dependencies = [
    "fastparquet>=2024.11.0",
    'geopandas>=1.1.1',
    'loguru',
    'pip',
    "pyarrow>=21.0.0",
]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[tool.ruff]
exclude = [
  '.bzr',
  '.direnv',
  '.eggs',
  '.git',
  '.git-rewrite',
  '.hg',
  '.ipynb_checkpoints',
  '.mypy_cache',
  '.nox',
  '.pants.d',
  '.pyenv',
  '.pytest_cache',
  '.pytype',
  '.ruff_cache',
  '.svn',
  '.tox',
  '.venv',
  '.vscode',
  '__pypackages__',
  '_build',
  'buck-out',
  'build',
  'dist',
  'node_modules',
  'site-packages',
  'venv',
]
line-length = 80
indent-width = 4
src = ['biodiv']
include = ['pyproject.toml', 'biodiv/**/*.py']

[tool.ruff.lint]
select = ['E4', 'E7', 'E9', 'F', 'I']
ignore = ['F403']
fixable = ['ALL']
unfixable = []
dummy-variable-rgx = '^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$'

[tool.ruff.lint.isort]
known-first-party = ['biodiv']
required-imports = ["from __future__ import annotations"]

[tool.ruff.format]
quote-style = 'single'
indent-style = 'space'
skip-magic-trailing-comma = false
line-ending = 'auto'

[dependency-groups]
dev = [
    "mypy>=1.17.0",
    "pytest>=8.4.1",
    "ruff>=0.12.5",
]
